refactor: SubClassContent 상위로 이동 및 로직 분리
